+++ pom.xml
@@ -3,7 +3,7 @@
   <groupId>com.github.dsipaint</groupId>
   <artifactId>AMGN</artifactId>
 
-  <version>1.0.2</version>
+  <version>1.1</version>
 
   <name>AMGN</name>
