+++ src/main/java/com/github/dsipaint/AMGN/entities/listeners/IListener.java
@@ -1,0 +1,4 @@
+package com.github.dsipaint.AMGN.entities.listeners;
+
+/* this is a placeholder interface to identify AMGN listeners (discord or otherwise) within the AMGN codebase */
+public interface IListener {}
