+++ src/main/java/com/github/dsipaint/AMGN/entities/listeners/managed/menu/Menu.java
@@ -1,4 +1,4 @@
-package com.github.dsipaint.AMGN.entities.listeners.menu;
+package com.github.dsipaint.AMGN.entities.listeners.managed.menu;
 
 import java.time.Instant;
 import java.util.concurrent.ScheduledFuture;
