+++ src/main/java/com/github/dsipaint/AMGN/entities/plugins/intrinsic/consistency/MenuDeleteListener.java
@@ -2,8 +2,8 @@ package com.github.dsipaint.AMGN.entities.plugins.intrinsic.consistency;
 
 import com.github.dsipaint.AMGN.AMGN;
 import com.github.dsipaint.AMGN.entities.GuildNetwork;
-import com.github.dsipaint.AMGN.entities.listeners.menu.Button;
-import com.github.dsipaint.AMGN.entities.listeners.menu.Menu;
+import com.github.dsipaint.AMGN.entities.listeners.managed.menu.Button;
+import com.github.dsipaint.AMGN.entities.listeners.managed.menu.Menu;
 
 import net.dv8tion.jda.api.events.message.MessageDeleteEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
