+++ src/main/java/com/github/dsipaint/AMGN/entities/plugins/intrinsic/help/HelpCommand.java
@@ -5,12 +5,12 @@ import java.util.function.Consumer;
 
 import com.github.dsipaint.AMGN.AMGN;
 import com.github.dsipaint.AMGN.entities.GuildNetwork;
-import com.github.dsipaint.AMGN.entities.listeners.Command;
 import com.github.dsipaint.AMGN.entities.listeners.CommandEvent;
-import com.github.dsipaint.AMGN.entities.listeners.DefaultCommand;
-import com.github.dsipaint.AMGN.entities.listeners.ListenerWrapper;
-import com.github.dsipaint.AMGN.entities.listeners.menu.ScrollMenuBuilder;
-import com.github.dsipaint.AMGN.entities.listeners.menu.MenuBuilder.InvalidMenuException;
+import com.github.dsipaint.AMGN.entities.listeners.managed.Command;
+import com.github.dsipaint.AMGN.entities.listeners.managed.DefaultCommand;
+import com.github.dsipaint.AMGN.entities.listeners.managed.ListenerWrapper;
+import com.github.dsipaint.AMGN.entities.listeners.managed.menu.ScrollMenuBuilder;
+import com.github.dsipaint.AMGN.entities.listeners.managed.menu.MenuBuilder.InvalidMenuException;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 
@@ -35,7 +35,7 @@ public final class HelpCommand implements Consumer<CommandEvent>
 
 			AMGN.plugin_listeners.forEach((plugin, list) ->
 			{
-				if(ListenerWrapper.pluginShouldRun(plugin.getName(), e.getGuild()))
+				if(ListenerWrapper.applyWhitelistBlacklistRules(plugin.getName(), e.getGuild()))
 				{
 					list.forEach(listener ->
 					{
