+++ src/main/java/com/github/dsipaint/AMGN/entities/plugins/intrinsic/permissions/ListPermissionsCommand.java
@@ -15,8 +15,8 @@ import org.yaml.snakeyaml.Yaml;
 import com.github.dsipaint.AMGN.AMGN;
 import com.github.dsipaint.AMGN.entities.GuildNetwork;
 import com.github.dsipaint.AMGN.entities.listeners.CommandEvent;
-import com.github.dsipaint.AMGN.entities.listeners.menu.MenuBuilder.InvalidMenuException;
-import com.github.dsipaint.AMGN.entities.listeners.menu.ScrollMenuBuilder;
+import com.github.dsipaint.AMGN.entities.listeners.managed.menu.ScrollMenuBuilder;
+import com.github.dsipaint.AMGN.entities.listeners.managed.menu.MenuBuilder.InvalidMenuException;
 import com.github.dsipaint.AMGN.entities.plugins.Plugin;
 import com.github.dsipaint.AMGN.io.IOHandler;
 import com.github.dsipaint.AMGN.io.Permissions;
